remove mock data from CourseContainer for lj details
