+++ app/src/components/learningJourney/learningJourneyDetails/CourseCard.jsx
